CellRangeSelector: add the onCellRangeSelectionOngoing event, which fires during the drag operation. It receives a parameter containing both the 'range' and 'currentCell' where 'range' is the currently collected UNORDERED range (range.start references the cell which started the selection; 'range.end' is either empty or references the previously hovered cell during the drag operation) and 'currentCell' references the currently hovered cell (which will be treated as the other selection corner).
The .currentCell.row/.cell values may be tweaked in the event handler: this will affect the selected range.

The event handler may return any truthy value to signal the selector that this cell should NOT be included in the selection.

This event is useful when you need to apply custom cell selection filters to the drag operation, e.g. restricting cell selections to single columns.