fix (if somewhat hacky) for an issue found in Chrome where the stylesheet which is generated at run-time by slickgrid was not parsed in time, i.e. was not available via the DOM a few JS statements later.
The idea for the fix is that the browser needs some air to breathe outside the current JS run, isetTimeout-triggered re-issue of the relevant slickgrid code should resolve the issue of the <style> not getting parsed. Slickgrid code investigation turns up that the style info is only really needed by the code which updates the column widths, so we can get away with executing the entire .init() code and then in said setTimeout() re-run the .init() until it reports utter completion to the caller.

Hence the user code now looks like this:

       _grid = new Slick.Grid("#myGrid", dataView, columns, options);

        // slickgrid returns TRUE when EVERYTHING was done; otherwise FALSE and we will need to give it a little air so it can parse styles....
        var slicktimer = false;
        var slick_tickler = function() {
            console.log("###### grid.init: did it find its own run-time generated stylesheet yet?");
            if (!_grid.init()) {
                // retry until we're good
                console.log("###### slickgrid did not completely INIT; retrying in a timer...");
                slicktimer = setTimeout(slick_tickler, 200);
            } else {
                console.log("###### slickgrid finished all initialization, including stylesheet-related work");
            }
        };
        slick_tickler();