reducing the explosion in function call parameters as we move from a notion where a cell carries a value (and an internal formatting) to a notion where the cell formatters (and editors) should have full access to the entire cell data/metadata set, i.e. they should be able to tweak the cell DIV node itself! -- this notion is developed as DOM work and DOM complexity is far more expensive at run-time than JavaScript code complexity.
This move is also strongly driven by the notion that we cannot store the DATA for large grids in a data 2D matrix specifically for slickgrid presentation: in real world large applications this acts as a 'store/cache' which must be kept up to date/synchronized and takes additional memory space. That can be very useful to have and use, but when you rely on DataView-based abstractions it can be very beneficial to forego this additional special-purpose store and obtain your slickgrid cell data straight from the horse's mouth i.e. the data provider which is referenced/linked by the DataView instance bound to your slickgrid. Particularly when that data source has already sophisticated internal precalculus and caching mechanisms in place -- yes, this implies the added overhead of at least one function call per grid cell data value access, but given the first notion/focusmove mentioned at the top, one can expect to need several data/metadata items per cell to be delivered through the DataView: then that function call (or two) per cell starts to look like minor stuff when we consider that such a call could deliver all data/metadata for the cell in one quick return.