Don't read and write width property in a loop. Instead get all widths first in a batch and then set all widths in a batch. Before this change the browser relayouted once for every single column. After this change the browser will relayout once if applyColumnWidths is called.
Conflicts:
	slick.grid.js

Note: I took care of this in a different way: by using the same .l/.r left/right style positioning approach for headers as was used for the cells in the canvas/viewport already *plus* always updating all those styles for any width change: this removes the need to 'measure' the width and thus removes the cost of a DOM tree re-layout implicit action by the browser.

'cherrypicked' to flag this commit as inspected & processed.