bugfix: fix issue where each resize action of any grid would produce one **dangling row**, i.e. after a while you'ld end up with a series of `slick-row-0` rows, all overlapping!
How to reproduce: run the examples/example-0086-40.grids.on.a.page.test.html test and drag the size slider to the left so that all data rows are completely done. Twiddle the slider to generate additional resize 'events', each one producing one more dangling '0' row, which is DOM cruft not accessible via slickgrid rowsCache any more!

This fix corrects this loss of control.